# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "DemoService",
        "description": "Demo service for the ConnectAPI marketplace and development cycle.",
        "contact": {
            "email": "ConnectAPI@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/test/query_test": {
            "get": {
                "tags": [
                    "test"
                ],
                "summary": "Query Test",
                "operationId": "query_test",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "A",
                            "type": "integer"
                        },
                        "name": "a",
                        "in": "query"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "B",
                            "type": "string"
                        },
                        "name": "b",
                        "in": "query"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "C",
                            "type": "boolean"
                        },
                        "name": "c",
                        "in": "query"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "D",
                            "type": "number"
                        },
                        "name": "d",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/query_optional_test": {
            "get": {
                "tags": [
                    "test"
                ],
                "summary": "Query Optional Test",
                "operationId": "query_optional_test",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "A",
                            "type": "integer",
                            "default": 0
                        },
                        "name": "a",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "B",
                            "type": "string",
                            "default": ""
                        },
                        "name": "b",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "C",
                            "type": "boolean",
                            "default": true
                        },
                        "name": "c",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "D",
                            "type": "number",
                            "default": 0.1
                        },
                        "name": "d",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/body_test": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "Body Test",
                "operationId": "body_test",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TestIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/body_optional_test": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "Body Optional Test",
                "operationId": "body_optional_test",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "Body",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/TestIn"
                                    }
                                ],
                                "default": {
                                    "a": 0,
                                    "a_o": 0,
                                    "b": "",
                                    "b_o": "",
                                    "c": true,
                                    "c_o": true,
                                    "d": 0.1,
                                    "d_o": 0.1,
                                    "e": {
                                        "a": 1
                                    },
                                    "e_o": {
                                        "a": 1
                                    },
                                    "f": [
                                        0
                                    ],
                                    "f_o": [
                                        0
                                    ],
                                    "g": {
                                        "a": 0
                                    },
                                    "g_o": {
                                        "a": 0
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/headers_test": {
            "get": {
                "tags": [
                    "test"
                ],
                "summary": "Headers Test",
                "operationId": "headers_test",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "A",
                            "type": "string"
                        },
                        "name": "a",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/headers_optional_test": {
            "get": {
                "tags": [
                    "test"
                ],
                "summary": "Headers Optional Test",
                "operationId": "headers_optional_test",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "A",
                            "type": "string",
                            "default": "hello"
                        },
                        "name": "a",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/form_test": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "Form Test",
                "operationId": "form_test",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/BodyForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/form_optional_test": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "Form Optional Test",
                "operationId": "form_optional_test",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/BodyFormOptional"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/cookies_test": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "Cookies Test",
                "operationId": "cookies_test",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "A",
                            "type": "string"
                        },
                        "name": "a",
                        "in": "cookie"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/cookies_optional_test": {
            "post": {
                "tags": [
                    "test"
                ],
                "summary": "Cookies Optional Test",
                "operationId": "cookies_optional_test",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "A",
                            "type": "string",
                            "default": "hello"
                        },
                        "name": "a",
                        "in": "cookie"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/file_test": {
            "put": {
                "tags": [
                    "test"
                ],
                "summary": "File Test",
                "operationId": "file_test",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BodyFile"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/file_optional_test": {
            "put": {
                "tags": [
                    "test"
                ],
                "summary": "File Optional Test",
                "operationId": "file_optional_test",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BodyFileOptional"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/test/path_test/{a}/{b}/{c}": {
            "put": {
                "tags": [
                    "test"
                ],
                "summary": "Path Test",
                "operationId": "path_test",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "A",
                            "type": "string"
                        },
                        "name": "a",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "B",
                            "type": "integer"
                        },
                        "name": "b",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "schema": {
                            "title": "C",
                            "type": "number"
                        },
                        "name": "c",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BodyFileOptional": {
                "title": "BodyFileOptional",
                "type": "object",
                "properties": {
                    "a": {
                        "title": "A",
                        "type": "string",
                        "format": "binary",
                        "default": "file_bytes"
                    }
                }
            },
            "BodyFile": {
                "title": "BodyFile",
                "required": [
                    "a"
                ],
                "type": "object",
                "properties": {
                    "a": {
                        "title": "A",
                        "type": "string",
                        "format": "binary"
                    }
                }
            },
            "BodyFormOptional": {
                "title": "BodyFormOptional",
                "type": "object",
                "properties": {
                    "a": {
                        "title": "A",
                        "type": "string",
                        "default": "hello"
                    }
                }
            },
            "BodyForm": {
                "title": "BodyForm",
                "required": [
                    "a"
                ],
                "type": "object",
                "properties": {
                    "a": {
                        "title": "A",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "Nested": {
                "title": "Nested",
                "type": "object",
                "properties": {
                    "a": {
                        "title": "A",
                        "type": "integer",
                        "default": 0
                    }
                }
            },
            "TestIn": {
                "title": "TestIn",
                "required": [
                    "a",
                    "b",
                    "c",
                    "d",
                    "e",
                    "f",
                    "g"
                ],
                "type": "object",
                "properties": {
                    "a": {
                        "title": "A",
                        "type": "integer"
                    },
                    "a_o": {
                        "title": "A O",
                        "type": "integer",
                        "default": 0
                    },
                    "b": {
                        "title": "B",
                        "type": "string"
                    },
                    "b_o": {
                        "title": "B O",
                        "type": "string",
                        "default": ""
                    },
                    "c": {
                        "title": "C",
                        "type": "boolean"
                    },
                    "c_o": {
                        "title": "C O",
                        "type": "boolean",
                        "default": true
                    },
                    "d": {
                        "title": "D",
                        "type": "number"
                    },
                    "d_o": {
                        "title": "D O",
                        "type": "number",
                        "default": 0.1
                    },
                    "e": {
                        "title": "E",
                        "type": "object"
                    },
                    "e_o": {
                        "title": "E O",
                        "type": "object",
                        "default": {
                            "a": 1
                        }
                    },
                    "f": {
                        "title": "F",
                        "type": "array",
                        "items": {}
                    },
                    "f_o": {
                        "title": "F O",
                        "type": "array",
                        "items": {},
                        "default": [
                            0
                        ]
                    },
                    "g": {
                        "$ref": "#/components/schemas/Nested"
                    },
                    "g_o": {
                        "title": "G O",
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Nested"
                            }
                        ],
                        "default": {
                            "a": 0
                        }
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}